Add lifecycle hooks for Worker as well (start, stop)
As these would be needed for job-iteration to know when the worker is
going to stop.
